agents-python/server/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.10.0
PyPDF2==3.0.1
boto3==1.28.62
botocore==1.31.62
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-embeddings==0.1.0
langchain-llms==0.1.0
langchain-vectorstores==0.1.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
